+++ kttk/folder_manager.py
@@ -70,13 +70,14 @@ def init_entity(entity_type, entity_id):
             f.write(content)
 
         # register the created paths in database
-        path_cache_manager.register_path(path, context)
+        path_cache_manager.register_path(os.path.dirname(file_path), context)
 
     # register entity folder
     entity_folder_template = template_manager.get_route_template(
         '{}_folder'.format(entity_type))
     entity_folder = template_manager.format_template(entity_folder_template, context_dict)
 
-    path_cache_manager.register_path(entity_folder, context)
+    if entity_folder != "":
+        path_cache_manager.register_path(entity_folder, context)
 
     # run setup hooks todo implement setup hooks