+++ kttk/path_cache_manager.py
@@ -8,14 +8,24 @@ class PathNotRegistered(Exception):
         super(PathNotRegistered, self).__init__("No Context registered for path {}".format(path))
 
 
+class InvalidPath(Exception):
+
+    def __init__(self, path):
+        super(InvalidPath, self).__init__("Path {} is not valid to be registered in database".format(path))
+
+
 def register_path(path, context):
     # type: (str, Context) -> dict
     """
     registeres given context for given path in database, so we can later get the context back from the path
-    :param path: path to register
+    :param path: path to register, None or "" not allowed!!!
     :param context: context to register
     :return: newly created path entry from database
     """
+    # check if path is valid
+    if not is_valid_path(path):
+        raise InvalidPath(path)
+
     # make path beautifull
     path = __good_path(path)
 
@@ -23,7 +33,7 @@ def register_path(path, context):
 
     path_entry_data = {}
     path_entry_data['path'] = path
-    path_entry_data['context'] = context.as_dict() # todo remove user information
+    path_entry_data['context'] = context.as_dict()  # todo remove user information
 
     return kt.create("path_entry", path_entry_data)
 
@@ -45,11 +55,10 @@ def unregister_path(path):
     entry_found = len(path_entries) > 0
 
     if entry_found:
-        path_entry = path_entries[0]
-
-        kt.delete('path_entry', path_entry['id'])
+        for path_entry in path_entries:
+            kt.delete('path_entry', path_entry['id'])
     else:
-        raise PathNotRegistered(path)
+        raise PathNotRegistered(path)  # todo better return boolean if path was unregistered
 
 
 def context_from_path(path):
@@ -71,7 +80,23 @@ def context_from_path(path):
         context = Context.from_dict(context_dicts[0]['context'])
         return context
     else:
-        raise PathNotRegistered(path) # todo better return None
+        raise PathNotRegistered(path)  # todo better return None
+
+
+def is_valid_path(path):
+    """
+    Checks if the path can be registered in the database.
+    Path is not allowed to be None or empty Strng
+    :param path: path to check
+    :return: True if path is valid, False otherwise
+    """
+    valid = False
+
+    if path:
+        if len(path) > 0:
+            valid = True
+
+    return valid
 
 
 def __good_path(path):