+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added task_presets_manager: Can now get task presets for given entity type from config
 - Ktrack Command: added support for task preset
 - Added type missing type hints to codebase
-- Added name sanitizer to remove all illegal chars from entity names 
+- Added name sanitizer to remove all illegal chars from entity names
+- Added project_bootstrapper to easily bootstrap a project for testing purposes and easy remove afterwards
 ## 0.1.0 - 2017-07-26
 ### Added
 - Tests: we can run tests for Maya